Intercept streaming responses on passthroughs
Introduces a `beforeResponseStream` option on passthroughs that
lets you intercept the stream before it is sent to the caller.
This allows you to duplicate the stream and write it to multiple
WriteableStreams in parallel, enabling you to capture the response
stream to a file in real-time as it streams back to the caller.